Fix problem where browser language does not have locale.
If you set your browser language to 'Arabic' in Chrome, the preferred language is
'ar', but the language picker was looking for the locale, e.g. wants 'ar-SA'.
Fix to check both locale and shortcode.
Add in rtl text direction to CSS for Arabic
Update README with instructions for adding to this.